refactor: move lock utils to models
