+++ packages/neuron-wallet/src/database/address/dao.ts
@@ -4,7 +4,7 @@ import { AddressType } from '../../models/keys/address'
 import { getConnection } from './ormconfig'
 import TransactionsService from '../../services/transactions'
 import CellsService from '../../services/cells'
-import LockUtils from '../../utils/lock-utils'
+import LockUtils from '../../models/lock-utils'
 
 export interface Address {
   walletId: string
