+++ packages/neuron-wallet/src/database/chain/meta-info.ts
@@ -1,5 +1,5 @@
 import FileService from '../../services/file'
-import { SystemScript } from '../../utils/lock-utils'
+import { SystemScript } from '../../models/lock-utils'
 
 const moduleName = 'cells'
 const fileName = 'meta-info.json'
