+++ packages/neuron-wallet/src/services/sync-blocks.ts
@@ -7,7 +7,7 @@ import TransactionsService from './transactions'
 import OutputEntity from '../database/chain/entities/output'
 import SyncInfoEntity from '../database/chain/entities/sync-info'
 import NodeService from './node'
-import LockUtils from '../utils/lock-utils'
+import LockUtils from '../models/lock-utils'
 import TypeConvert from '../types/type-convert'
 import { NetworkWithID } from './networks'
 import AddressesUsedSubject from '../models/subjects/addresses-used-subject'
