+++ packages/neuron-wallet/src/services/wallets.ts
@@ -7,7 +7,7 @@ import Keystore from '../models/keys/keystore'
 import Store from '../models/store'
 import NodeService from './node'
 import FileService from './file'
-import LockUtils from '../utils/lock-utils'
+import LockUtils from '../models/lock-utils'
 import windowManager from '../models/window-manager'
 import { Channel, ResponseCode } from '../utils/const'
 import { Witness, TransactionWithoutHash, Input } from '../types/cell-types'
