+++ packages/neuron-wallet/src/startup/sync-block-task/task.ts
@@ -2,7 +2,7 @@ import { remote } from 'electron'
 import { Subject } from 'rxjs'
 import { initConnection as initAddressConnection } from '../../database/address/ormconfig'
 import AddressService from '../../services/addresses'
-import LockUtils from '../../utils/lock-utils'
+import LockUtils from '../../models/lock-utils'
 import AddressesUsedSubject from '../../models/subjects/addresses-used-subject'
 import BlockListener from '../../services/sync/block-listener'
 import { NetworkWithID } from '../../services/networks'
